@@ -1,13 +1,12 @@
 name: ondato_flutter
 description: A new flutter plugin project.
-version: 0.0.2
+version: 0.1.0
 author:
 homepage:
 repository: "https://github.com/ondato/sample-sdk-app-flutter"
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
