@@ -1,21 +1,19 @@
 import 'dart:ui';
 
-import 'package:flutter/foundation.dart';
-
 enum OndatoEnvironment { test, live }
 enum OndatoLanguage { en, de, lt, lv, et, ru }
 
 extension OndatoEnvironmentExt on OndatoEnvironment {
-  String toMap() => this.toString().split('.')?.elementAt(1);
+  String? toMap() => this.toString().split('.').elementAt(1);
 }
 
 extension OndatoLanguageExt on OndatoLanguage {
-  String toMap() => this.toString().split('.')?.elementAt(1);
+  String? toMap() => this.toString().split('.').elementAt(1);
 }
 
 class OndatoServiceConfiguration {
-  OndatoIosAppearance appearance;
-  OndatoFlowConfiguration flowConfiguration;
+  OndatoIosAppearance? appearance;
+  OndatoFlowConfiguration? flowConfiguration;
   OndatoEnvironment mode;
   OndatoLanguage language;
   OndataCredencials credentials;
@@ -25,27 +23,25 @@ class OndatoServiceConfiguration {
     this.flowConfiguration,
     this.mode = OndatoEnvironment.test,
     this.language = OndatoLanguage.en,
-    this.credentials,
-  }) : assert(credentials != null, 'Credencials must be provided');
+    required this.credentials,
+  });
 
   Map<String, dynamic> toMap() {
     return {
       'appearance': appearance?.toMap(),
       'flowConfiguration': flowConfiguration?.toMap(),
-      'mode': mode?.toMap(),
-      'language': language?.toMap(),
-      'credentials': credentials?.toMap(),
+      'mode': mode.toMap(),
+      'language': language.toMap(),
+      'credentials': credentials.toMap(),
     };
   }
 }
 
 class OndataCredencials {
   OndataCredencials({
-    @required this.accessToken,
-    @required this.identificationId,
-  })  : assert(
-            accessToken != null, 'Use or username and password or accessToken'),
-        assert(identificationId != null, 'IdentificationID must not be null');
+    required this.accessToken,
+    required this.identificationId,
+  });
   final String accessToken;
   final String identificationId;
 
@@ -56,28 +52,28 @@ class OndataCredencials {
 
 class OndatoFlowConfiguration {
   /// Should the splash screen be shown
-  bool showSplashScreen;
+  bool? showSplashScreen;
 
   /// Should the start screen be shown
-  bool showStartScreen;
+  bool? showStartScreen;
 
   /// Should the consent screen be shown
-  bool showConsentScreen;
+  bool? showConsentScreen;
 
   /// Should a selfie with document be requested when taking document pictures
-  bool showSelfieAndDocumentScreen;
+  bool? showSelfieAndDocumentScreen;
 
   /// Should the success window be shown
-  bool showSuccessWindow;
+  bool? showSuccessWindow;
 
   /// Allows user to skip liveness check in case of failure
-  bool ignoreLivenessErrors;
+  bool? ignoreLivenessErrors;
 
   /// Allows user to skip document verification error result checks
-  bool ignoreVerificationErrors;
+  bool? ignoreVerificationErrors;
 
   /// Should the verification process be recorded
-  bool recordProcess;
+  bool? recordProcess;
 
   OndatoFlowConfiguration({
     this.showSplashScreen = true,
@@ -104,7 +100,7 @@ class OndatoFlowConfiguration {
   }
 
   factory OndatoFlowConfiguration.fromMap(Map<String, dynamic> map) {
-    if (map == null) return null;
+    // if (map == null) return null;
 
     return OndatoFlowConfiguration(
       showSplashScreen: map['showSplashScreen'],
@@ -121,34 +117,34 @@ class OndatoFlowConfiguration {
 
 class OndatoIosAppearance {
   /// Logo image that can be shown in the splash screen
-  String logoImageBase64;
+  String? logoImageBase64;
 
   /// background color of the `ProgressBarView` which guides the user through the flow
-  Color progressColor;
+  Color? progressColor;
 
   /// background color of the primary action buttons
-  Color buttonColor;
+  Color? buttonColor;
 
   /// background color of the primary action buttons text
-  Color buttonTextColor;
+  Color? buttonTextColor;
 
   /// background color of the error message background
-  Color errorColor;
+  Color? errorColor;
 
   /// background color of the error message text color
-  Color errorTextColor;
+  Color? errorTextColor;
 
   /// regular text font
-  String regularFontName;
+  String? regularFontName;
 
   /// medium text font
-  String mediumFontName;
+  String? mediumFontName;
 
-  Color headerColor;
+  Color? headerColor;
 
-  Color acceptButtonColor;
+  Color? acceptButtonColor;
 
-  Color declineButtonColor;
+  Color? declineButtonColor;
 
   OndatoIosAppearance({
     this.logoImageBase64,
@@ -185,8 +181,8 @@ class OndatoIosAppearance {
 }
 
 class OndatoException implements Exception {
-  OndatoError error;
-  String identificationId;
+  OndatoError? error;
+  String? identificationId;
   OndatoException(this.identificationId, error) {
     switch (error) {
       case 'cancelled':
